Make sure we reap SSH process on Close()
System transport can get stuck on the Read call because of
pty implementation details (see https://github.com/scrapli/scrapligo/issues/174)
for more details.

This makes sure that we kill the SSH process when the transport
is closed, and in turn makes sure that the rest of the components (channel, drivers)
can shutdown properly.

I also added a test simulating a blocking (stuck) SSH server
and a non blokcing SSH server (non blocking showed that it's fine
to kill an already defunct process).

Signed-off-by: Francois Espinet <francois.espinet@corp.ovh.com>